Use NewCmd* pattern for describe and describe service.
